requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==46.0.1
distro==1.9.0
Django==5.2.6
django-filter==25.2
django-jazzmin==3.0.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
mysqlclient==2.2.7
openai==2.0.1
packaging==25.0
pillow==11.3.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyMySQL==1.1.2
PyPDF2==3.0.1
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
